+++ kilometer-backend-backoffice/src/main/java/com/kilometer/backend/security/CustomUserDetailsService.java
@@ -1,0 +1,31 @@
+package com.kilometer.backend.security;
+
+import com.kilometer.domain.account.Account;
+import com.kilometer.domain.account.AccountService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final AccountService accountService;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Account account = accountService.findByUsername(username);
+        if (account == null) {
+            throw new UsernameNotFoundException(username);
+        }
+
+        return User.builder()
+                .username(account.getUsername())
+                .password(account.getPassword())
+                .roles(account.getRole())
+                .build();
+    }
+}
