+++ kilometer-backend-domain/build.gradle
@@ -11,7 +11,6 @@ dependencies {
     implementation "org.flywaydb:flyway-mysql:8.5.10"
     runtimeOnly 'org.mariadb.jdbc:mariadb-java-client'
     implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
 }
 
 bootRun.enabled = false
