+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/account/AccountRequest.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Data
 @Builder
@@ -24,7 +23,7 @@ public class AccountRequest {
                 .build();
     }
 
-    public void encodePassword(PasswordEncoder passwordEncoder) {
-        this.password = passwordEncoder.encode(this.password);
+    public void encodePassword (String password) {
+        this.password = password;
     }
 }
