+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/account/AccountService.java
@@ -1,36 +1,19 @@
 package com.kilometer.domain.account;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class AccountService implements UserDetailsService {
+public class AccountService {
 
     private final AccountRepository accountRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Account account = accountRepository.findByUsername(username);
-        if (account == null) {
-            throw new UsernameNotFoundException(username);
-        }
-
-        return User.builder()
-                .username(account.getUsername())
-                .password(account.getPassword())
-                .roles(account.getRole())
-                .build();
-    }
 
     public Account save(AccountRequest accountRequest) {
-        accountRequest.encodePassword(passwordEncoder);
         return accountRepository.save(accountRequest.makeAccount());
+    }
+
+    public Account findByUsername(String username) {
+        return accountRepository.findByUsername(username);
     }
 }
